Add files via upload
In this example, we have a directory called geometry that contains two files: __init__.py and geometry.py. The __init__.py file is empty, but it indicates that the geometry directory is a package. The geometry.py file contains the implementation of the functions for calculating areas of different shapes.